move gene network stuff to new file
